+++ metadata.rb
@@ -4,8 +4,9 @@ maintainer_email 'ejimenez@redborder.com'
 license          'All rights reserved'
 description      'Installs/Configures redborder manager'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.0.13'
+version          '0.0.14'
 
 depends 'zookeeper'
 depends 'kafka'
 depends 'druid'
+depends 'http2k'